requirements.txt
pytest==7.4.3
black==23.11.0
flake8==6.1.0
Flask==2.0.1
Werkzeug==2.0.3
python-dotenv==0.19.0
google-api-python-client==2.86.0
python-dateutil==2.9.0
pandas==2.2.1
pytz==2023.3
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
gunicorn==21.2.0
requests==2.31.0 